+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/controllers/operational_space.py
@@ -177,6 +177,10 @@ class OperationSpaceController:
     Operations.
     """
 
+    def initialize(self):
+        """Initialize the internals."""
+        pass
+
     def reset_idx(self, robot_ids: torch.Tensor = None):
         """Reset the internals."""
         pass